Added configurable interface state.
This is usefull if main interface doesn't have ip, but has vlan that has
IP. Before this change vlan would be configure, but wouldn't work,
because physical interface would be down after reboot.

state option isn't required, but can be helpful.

However thic changes bridge creation a bit. Before this change bridge
would be created and would be up, however now sate needs to be set to 'up'
explicitly for bridge to work.

Another scenario is when you create vlan without IP, that will be
member of bridge to which jail will connect via e-pair.
In this case vlan interface also needs to be set up, or network won't
work for jail.